user-service/requirements.txt
fastapi
uvicorn
sqlalchemy
alembic
psycopg2-binary
pydantic
python-dotenv
passlib[bcrypt]
python-jose[cryptography]
requests
pydantic-settings